Voting works on the frontend
There is no limit on the number of times they can vote at the moment.
However, we can submit a vote from the frontend, store it on the back,
and modify the rating.

Running `elm-format` changed a good number of files, but major additions
to the frontend were made in `src/Pages/Text/Search.elm`. Note that
we've also added the `Vote` module.