Variable naming and backend seems to work
Looks like a student can only vote once, and naming changed to be
consistent with the term "vote". Next up is to handle data now being
passed to the frontend. The data being their previous "vote" if one
exists. On the front we need to disable buttons and change colors
accordingly.