Works from from to back
Caught the small bug on the backend, so now when the model comes back
with an updated list of texts, it'll be identical to what's already
shown on the page. So the rating returned matches what is calculated on
the frontend.

We return all of the texts by simply calling the `GET` endpoint on the
backend. It's equipped to handle the querystring that decides which
texts are shown in the search. This _could_ be broken out into another
function.