Made rating exclusive
You can only rate text that is either in progress or completed. Stylized
the buttons on the frontend to prevent clicking if not read. On the
backend we search the progress views to see if they've made an entry. If
not, they're not allowed to vote. They must have gotten there via
nefarious means!