Server response only
Cuts all the funny business on the frontend and just updates the vote
once the server responds. This causes an initial delay from when the
vote button is first pushed, but removes the bug where the vote switches
indecisively between the frontend's calculated value and the backend's
freshly updated value.