Text rating fixed
This fix makes the message returned from the server considerably
smaller. We've got three fields for one text rather than a completely
updated list of texts. Instead of changing the rating on the frontend,
and the vote, we wait until the server replies. At that point we update
the text. The bug created by balancing the frontend with the backend
rating is now gone; the solution is to depend solely on the backend for
this task.