Updated student guide
Noticed that the student guide was reusing an image in the wrong place.
Decided to take a new screenshot and add it to the settings page. Also
moved all of the images used in the guide into a more specific directory
`public/img/tutorial/student/` which gives us space for separate images
used in the content creator guide.