Backend handles dashboard connected flag complete
Updated the form for dashboard_connected, was hung up on a bug for a
long time (needed to exclude the new field in `user_student`). Wired up
entirely with APIView for the endpoint
`api/student/<int:pk>/connect_to_dashboard`. Sending
`{"connect_to_dashboard":<bool>}` will make an entry in the table
`user_studentdashboarduser` and `user_studentdashboarduserrange`. It'll
take that `id` and place it in the `user_student` table under
`dashboard_user_id` as well.