Sync on login
The backend has been largely flushed out to sync texts on login. We
don't have the exact message to pass to the dashboard just yet, but it's
been carved out.

Configuration variables have been placed in settings.py so that we can
change things like endpoints and frequency of data being pushed.