Check box for connecting to Dashboard
This commit brings in the check box on the Student Profile Page which
confirms their interest in sharing data to the Flagship Connect
dashboard. When they log in it'll check to see the last time there was a
push of data. This interval can be set in `settings.py` as mentioned
in a previous commit.