Table statically rendered
Here is a partially complete version of the My Words table delivered as
a HTML page from the backend. This is less preferred for a variety of
reasons, but at the moment it's functional. Many of the functions may be
reused to deliver data to the SPA page (the next iteration of this feature).