First Time Completion table now uses date times
The table now looks just like the completion side, except the two right
most columns are (number of correct / total possible answers) and
percentage correct.

Modifications to the database are necessary for this to work. There will
need to be an `end_dt` column added to the
first_time_correct_firsttimecorrect table. That same column will need to
be added to the report_first_time_correct view.

To test this, acquire or create a compatible database, then read a text
you haven't read before. The metrics should show up in the First Time
Correct table on the student profile page.