First Time Completion table now uses date times (#432)
* First Time Completion table now uses date times

The table now looks just like the completion side, except the two right
most columns are (number of correct / total possible answers) and
percentage correct.

Modifications to the database are necessary for this to work. There will
need to be an `end_dt` column added to the
first_time_correct_firsttimecorrect table. That same column will need to
be added to the report_first_time_correct view.

To test this, acquire or create a compatible database, then read a text
you haven't read before. The metrics should show up in the First Time
Correct table on the student profile page.

* Filter bug fixed

Needed to filter on text difficulty when looping over various
difficulties and inserting those metrics into the performance report
JSON blob.

* adding migration files to hopefully statisfy tests

* Performance report pdf

Updated the PDF to reflect changes in the First Time Comprehension
table.

* Removed CSV link and cleaned commment

Had a bunch of commented templating code- removed it. CSV option removed
per request of project owner.

* Elm format ran on src/

* Removed unnecesary comments