Feature/my words table spa (#429)
* My Words phrase -> lemma

It makes more sense to use the lemma when saving to My Words. The phrase
itself is used in the sentence included in the My Words flashcard, but
the standalone word is now the lemma.

* your words -> My Words

Updated the links on the profile page to be more consistent

* Table statically rendered

Here is a partially complete version of the My Words table delivered as
a HTML page from the backend. This is less preferred for a variety of
reasons, but at the moment it's functional. Many of the functions may be
reused to deliver data to the SPA page (the next iteration of this feature).

* WIP

Not sure how to get the API call to fetch the words upon initialization

* Plumbed from to back for table

We've got data coming from the backend to the front without issue.
my-words now sits at the top level under `Pages/`
Improvements to make:
coalesce the two decoders
stylize the table
put a link in the navbar

* CSS on My Words and cleaned

changed `#title` -> `guide-title` throughout
added My Words to the navbar
stylized the text on My Words page

* fixing typo

* Fixed broken tests, removed server side table

Fixed about 16 tests (improved some that were flaky)
 - zero out Demo Text rating (it must be between -1 and 1!)
 - force choose the next and prev button in hints modal
 - fresh baseline images
Removed server side table and plumbing

* removing Debug stmt

* Cypress tests

Adding cypress tests for the Guide page and the My Words page. Very
slight modifications to student_profile.spec.js

* Fake data

Adding an example to the My Words table if the student hasn't added any
My Words. It just says "example" in various ways. The table still
stretches to the width of its content. Hopefully that's fine?

* Elm format ran on src/

* Fixed namespacing error

* removed dead code (statically created my-words

* Removed comments, typo

Had a renegade typo, deleted dead code, and removed unnecessary
comments.