Devel without guide (#436)
* First Time Completion table now uses date times (#432)

* First Time Completion table now uses date times

The table now looks just like the completion side, except the two right
most columns are (number of correct / total possible answers) and
percentage correct.

Modifications to the database are necessary for this to work. There will
need to be an `end_dt` column added to the
first_time_correct_firsttimecorrect table. That same column will need to
be added to the report_first_time_correct view.

To test this, acquire or create a compatible database, then read a text
you haven't read before. The metrics should show up in the First Time
Correct table on the student profile page.

* Filter bug fixed

Needed to filter on text difficulty when looping over various
difficulties and inserting those metrics into the performance report
JSON blob.

* adding migration files to hopefully statisfy tests

* Performance report pdf

Updated the PDF to reflect changes in the First Time Comprehension
table.

* Removed CSV link and cleaned commment

Had a bunch of commented templating code- removed it. CSV option removed
per request of project owner.

* Elm format ran on src/

* Removed unnecesary comments

* Feature/my words table spa (#429)

* My Words phrase -> lemma

It makes more sense to use the lemma when saving to My Words. The phrase
itself is used in the sentence included in the My Words flashcard, but
the standalone word is now the lemma.

* your words -> My Words

Updated the links on the profile page to be more consistent

* Table statically rendered

Here is a partially complete version of the My Words table delivered as
a HTML page from the backend. This is less preferred for a variety of
reasons, but at the moment it's functional. Many of the functions may be
reused to deliver data to the SPA page (the next iteration of this feature).

* WIP

Not sure how to get the API call to fetch the words upon initialization

* Plumbed from to back for table

We've got data coming from the backend to the front without issue.
my-words now sits at the top level under `Pages/`
Improvements to make:
coalesce the two decoders
stylize the table
put a link in the navbar

* CSS on My Words and cleaned

changed `#title` -> `guide-title` throughout
added My Words to the navbar
stylized the text on My Words page

* fixing typo

* Fixed broken tests, removed server side table

Fixed about 16 tests (improved some that were flaky)
 - zero out Demo Text rating (it must be between -1 and 1!)
 - force choose the next and prev button in hints modal
 - fresh baseline images
Removed server side table and plumbing

* removing Debug stmt

* Cypress tests

Adding cypress tests for the Guide page and the My Words page. Very
slight modifications to student_profile.spec.js

* Fake data

Adding an example to the My Words table if the student hasn't added any
My Words. It just says "example" in various ways. The table still
stretches to the width of its content. Hopefully that's fine?

* Elm format ran on src/

* Fixed namespacing error

* removed dead code (statically created my-words

* Removed comments, typo

Had a renegade typo, deleted dead code, and removed unnecessary
comments.

* My Words phrase -> lemma (#423)

* My Words phrase -> lemma

It makes more sense to use the lemma when saving to My Words. The phrase
itself is used in the sentence included in the My Words flashcard, but
the standalone word is now the lemma.

* your words -> My Words

Updated the links on the profile page to be more consistent

* Fixing merge conflict

* Updating images in Student Guide (#422)

This updates bring in the checkbox for Flagship Connect as well as the
rating system on texts.

* Instructor profile tests (#419)

Basic cypress tests for the instructor profile page

* Bugfix/dependabot (#420)

* Dependabot recommendations

Looking to solve the issues that have been recommended to be fixed by
dependabot.

* Updating package tarball location

* Force deleting migrations

Branches were in conflict with one another, couldn't find the place to
remove all migrations. This needed to be done so we could overhaul the
table with different column names and new columns.

* updating version numbers