Migrations
These migrations were necessary to make the database work with the newly
defined FTC model. Some of them are duplicates that were removed in the
previous commit. At the time, that seemed to be the way to do things.

To get the fifth and most important (read "recent") migration I had to
edit the fields by hand. They RenameField (!) and AddField as necessary.